+++ resources/admin/__init__.py
@@ -30,7 +30,7 @@ from ..models import (
     ReservationHomeMunicipalityField, ReservationHomeMunicipalitySet, Resource, ResourceTag, ResourceAccessibility,
     ResourceEquipment, ResourceGroup, ResourceImage, ResourceType, TermsOfUse,
     Unit, UnitAuthorization, UnitIdentifier, UnitGroup, UnitGroupAuthorization,
-    MaintenanceMessage, UniversalFormFieldType, ResourceUniversalField, ResourceUniversalFormOption, MaintenanceMode
+    UniversalFormFieldType, ResourceUniversalField, ResourceUniversalFormOption
 )
 from ..models.utils import generate_id
 from munigeo.models import Municipality
@@ -530,41 +530,6 @@ class RespaTokenAdmin(admin.ModelAdmin):
     raw_id_fields = ('user',)
 
 
-class MaintenanceModeInline(admin.TabularInline):
-    model = MaintenanceMode
-    fields = ('start', 'end', )
-    verbose_name = _('maintenance mode')
-    verbose_name_plural = _('maintenance modes')
-    extra = 0
-
-
-class MaintenanceMessageAdminForm(forms.ModelForm):
-    class Meta:
-        model = MaintenanceMessage
-        fields = ('start', 'end', 'message', )
-
-    def clean(self):
-        start = self.cleaned_data['start']
-        end = self.cleaned_data['end']
-        query = Q(end__gt=start, start__lt=end)
-        if self.instance and self.instance.pk:
-            query &= ~Q(pk=self.instance.pk)
-        collision = MaintenanceMessage.objects.filter(query)
-        if collision.exists():
-            raise ValidationError(_('maintenance message already exists.'))
-
-class MaintenanceMessageAdmin(TranslationAdmin):
-    form = MaintenanceMessageAdminForm
-    inlines = ( MaintenanceModeInline, )
-    fieldsets = (
-        (_('General'), {
-            'fields': (
-                'start',
-                'end',
-                'message'
-            ),
-        }),
-    )
 
 admin_site.register(ResourceImage, ResourceImageAdmin)
 admin_site.register(Resource, ResourceAdmin)
@@ -594,6 +559,5 @@ admin.site.register(ResourceAccessibility, ResourceAccessibilityAdmin)
 if admin.site.is_registered(Token):
     admin.site.unregister(Token)
 admin_site.register(Token, RespaTokenAdmin)
-admin_site.register(MaintenanceMessage, MaintenanceMessageAdmin)
 admin_site.register(UniversalFormFieldType, UniversalFieldAdmin)
 admin_site.register(ResourceUniversalFormOption, ResourceUniversalFormOptionAdmin)
