+++ resources/api/resource.py
@@ -45,7 +45,7 @@ from resources.models import (
     ResourceImage, ResourceType, ResourceEquipment, TermsOfUse, Equipment, ReservationMetadataSet,
     ReservationMetadataField, ReservationHomeMunicipalityField,
     ReservationHomeMunicipalitySet, ResourceDailyOpeningHours, UnitAccessibility, Unit, ResourceTag,
-    ResourceUniversalField, ResourceUniversalFormOption, UniversalFormFieldType, MaintenanceMode
+    ResourceUniversalField, ResourceUniversalFormOption, UniversalFormFieldType
 )
 from resources.models.resource import determine_hours_time_range
 from payments.models import Product
@@ -65,6 +65,7 @@ from .resource_field import UniversalFormFieldTypeSerializer
 from rest_framework.settings import api_settings as drf_settings
 from rest_framework.relations import PrimaryKeyRelatedField
 from resources.models.utils import log_entry
+from maintenance.models import MaintenanceMode
 
 from drf_yasg import openapi
 from drf_yasg.utils import swagger_auto_schema
@@ -656,8 +657,6 @@ class ResourceSerializer(ExtraDataMixin, TranslatedModelSerializer, munigeo_api.
             [tag.label for tag in ResourceTag.objects.filter(resource=obj)] + list(obj.tags.names())))
 
     def get_user_permissions(self, obj):
-        if MaintenanceMode.objects.active().exists():
-            return { }
         request = self.context.get('request', None)
         prefetched_user = self.context.get('prefetched_user', None)
 
@@ -665,7 +664,8 @@ class ResourceSerializer(ExtraDataMixin, TranslatedModelSerializer, munigeo_api.
             user = prefetched_user or request.user
 
         can_make_reservations_for_customers = obj.can_create_reservations_for_other_users(user) if request else False
-        return {
+
+        permissions = {
             'can_make_reservations': obj.can_make_reservations(user) if request else False,
             **({'can_make_reservations_for_customer': can_make_reservations_for_customers} if (request and can_make_reservations_for_customers) else {}),
             'can_ignore_opening_hours': obj.can_ignore_opening_hours(user) if request else False,
@@ -674,6 +674,12 @@ class ResourceSerializer(ExtraDataMixin, TranslatedModelSerializer, munigeo_api.
             'is_viewer': obj.is_viewer(user) if request else False,
             'can_bypass_payment': obj.can_bypass_payment(user) if request else False,
         }
+
+
+        if MaintenanceMode.objects.active().exists():
+            return permissions.fromkeys(permissions, False)
+
+        return permissions
 
     def get_is_favorite(self, obj):
         request = self.context.get('request', None)
@@ -794,9 +800,6 @@ class ResourceSerializer(ExtraDataMixin, TranslatedModelSerializer, munigeo_api.
             self.context.update(times)
 
     def get_opening_hours(self, obj):
-        if MaintenanceMode.objects.active().exists():
-            return [ ]
-
         if 'start' in self.context:
             start = self.context['start']
             end = self.context['end']
