+++ resources/translation.py
@@ -5,7 +5,7 @@ from .models import (
     Purpose, Resource, ResourceEquipment,
     ResourceImage, ResourceType, TermsOfUse, Unit,
     UnitGroup, Reservation, ReservationHomeMunicipalityField,
-    MaintenanceMessage, UniversalFormFieldType, ResourceUniversalField,
+    UniversalFormFieldType, ResourceUniversalField,
     ResourceUniversalFormOption,
 )
 
@@ -74,12 +74,6 @@ class AccessibilityViewpointTranslationOptions(TranslationOptions):
 class ReservationHomeMunicipalityFieldTranslationOptions(TranslationOptions):
     fields = ('name',)
     required_languages = ('fi', 'en', 'sv')
-
-@register(MaintenanceMessage)
-class MaintenanceMessageTranslationOptions(TranslationOptions):
-    fields = ('message', )
-    required_languages = ('fi', 'en', 'sv', )
-
 
 @register(UniversalFormFieldType)
 class UniversalFormFieldTranslationOptions(TranslationOptions):
